+++ app/models/digitization/newspaper.rb
@@ -4,6 +4,8 @@ class Digitization::Newspaper < JupiterCore::Depositable
 
   belongs_to :owner, class_name: 'User'
 
+  has_solr_exporter Exporters::Solr::Digitization::BookExporter
+
   validates :publication_code, uniqueness: { scope: [:year, :month, :day] }
   validates :year, :month, :day, presence: true, if: :publication_code?