+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New entries in this file should aim to provide a meaningful amount of informatio
 
 - Jupiter II work is continuing to incorporate Digitized materials into Jupiter in the digitalcollections namespace.
   - newspaper metadata for ACN digitization [#2645](https://github.com/ualbertalib/jupiter/issues/2645)
+  - newspaper show page [#2649](https://github.com/ualbertalib/jupiter/issues/2649)
   
 ## [2.3.1] - 2021-12-07