+++ CHANGELOG.md
@@ -28,8 +28,8 @@ New entries in this file should aim to provide a meaningful amount of informatio
 - We had a request by a researcher to attach several 2.8 Gb zip files to an existing object.  We couldn't fulfill this request because of the way we were storing metadata about the file.  Using `Integer` put an artificial limitation of 2,147,483,647 (2^31-1) on the size of files we could attach.  We migrate the `byte_size` of blobs to use `BigInt`, 9,223,372,036,854,775,807 (2^63-1), instead.
 
 - Jupiter II work is continuing to incorporate Digitized materials into Jupiter in the digitalcollections namespace.
-  - Add task that will kick off setup job for attaching artifacts to digitization objects [#2011](https://github.com/ualbertalib/jupiter/issues/2011)
-  - Add job that will attach artifacts (high resolution pdf, full text, and link to preservation storage) to digitization objects [#2011](https://github.com/ualbertalib/jupiter/issues/2011)
+  - Add task that will kick off job for attaching artifacts (high resolution pdf, full text, and link to preservation storage) to digitization objects [#2011](https://github.com/ualbertalib/jupiter/issues/2011)
+
 ### Chores
 - bump sidekiq-unique-jobs from 7.0.12 to 7.1.8 and fix a long missed deprecation
 - fixed deprecation warning on tests [#2604](https://github.com/ualbertalib/jupiter/issues/2604)